fix bug in connectivity optimization
Due to the linearized channel model, the optimization may drive the system
infeasible at the next iteration. And, due to the stopping criterion, it may be
that the system does not stop there, driving the system into a undesirable
state. To fix this issue, check whether the solution is near infeasibility at
the end of the optimization step and decide whether to update the comm agent
configuration accordingly.

Also add a verbose mode to help with easily toggling print statements useful for
determining the behavior of the optimizer.